test svg map

+++ src/sql/install/papervote_optical.sql
@@ -2,7 +2,7 @@ DELIMITER ;
 
 create table if not exists `papervote_optical` (
 
-  `pagination_id` integer not null,
+  `pagination_id` bigint not null,
 
   `box_id` varchar(32) not null,
   `stack_id` varchar(32) not null,
@@ -18,7 +18,7 @@ create table if not exists `papervote_optical` (
 
 create table if not exists `papervote_optical_data` (
 
-  `pagination_id` integer not null,
+  `pagination_id` bigint not null,
     `data` longtext not null,
     primary key (`pagination_id`),
     constraint `fk_papervote_optical_data_pagination_id` foreign key (`pagination_id`) references `papervote_optical` (`pagination_id`) on delete cascade on update cascade
